+++ app/scripts/fhir/R4/lformsFHIR.js
@@ -19611,65 +19611,37 @@ var self = {
 
       if (dataType === 'CWE' || dataType === 'CNE') {
         var codeSystem = null,
-            coding = null;
+            coding = null; // item.defaultAnswer could be an array of multiple default values or a single value
 
-        if (this._answerRepeats(item) && Array.isArray(item.defaultAnswer)) {
-          // defaultAnswer has multiple values
-          for (var i = 0, iLen = item.defaultAnswer.length; i < iLen; i++) {
-            if (_typeof(item.defaultAnswer[i]) === 'object') {
-              coding = {
-                "code": item.defaultAnswer[i].code
-              };
+        var defaultAnswers = this._answerRepeats(item) && Array.isArray(item.defaultAnswer) ? item.defaultAnswer : [item.defaultAnswer]; // go through each default value, which could be a code object or a string
 
-              if (item.defaultAnswer[i].text !== undefined) {
-                coding.display = item.defaultAnswer[i].text;
-              } // code system
+        for (var i = 0, iLen = defaultAnswers.length; i < iLen; i++) {
+          if (_typeof(defaultAnswers[i]) === 'object') {
+            coding = {
+              "code": defaultAnswers[i].code
+            };
 
-
-              codeSystem = item.defaultAnswer[i].codeSystem || item.answerCodeSystem;
-
-              if (codeSystem) {
-                coding.system = LForms.Util.getCodeSystem(codeSystem);
-              }
-
-              answer = {};
-              answer[valueKey] = coding;
-              targetItem.initial.push(answer);
-            } // user typed answer that is not on the answer list.
-            else if (typeof item.defaultAnswer[i] === 'string') {
-                targetItem.initial.push({
-                  "valueString": item.defaultAnswer[i]
-                });
-              }
-          }
-        } // single selection, item.defaultAnswer is not an array
-        else {
-            if (_typeof(item.defaultAnswer) === 'object') {
-              coding = {
-                "code": item.defaultAnswer.code
-              };
-
-              if (item.defaultAnswer.text !== undefined) {
-                coding.display = item.defaultAnswer.text;
-              } // code system
+            if (defaultAnswers[i].text !== undefined) {
+              coding.display = defaultAnswers[i].text;
+            } // code system
 
 
-              codeSystem = item.defaultAnswer.codeSystem || item.answerCodeSystem;
+            codeSystem = defaultAnswers[i].codeSystem || item.answerCodeSystem;
 
-              if (codeSystem) {
-                coding.system = LForms.Util.getCodeSystem(codeSystem);
-              }
+            if (codeSystem) {
+              coding.system = LForms.Util.getCodeSystem(codeSystem);
+            }
 
-              answer = {};
-              answer[valueKey] = coding;
-              targetItem.initial.push(answer);
-            } // user typed answer that is not on the answer list.
-            else if (typeof item.defaultAnswer === 'string') {
-                targetItem.initial.push({
-                  "valueString": item.defaultAnswer
-                });
-              }
-          }
+            answer = {};
+            answer[valueKey] = coding;
+            targetItem.initial.push(answer);
+          } // user typed answer that is not on the answer list.
+          else if (typeof defaultAnswers[i] === 'string') {
+              targetItem.initial.push({
+                "valueString": defaultAnswers[i]
+              });
+            }
+        }
       } // for Quantity,
       // [{
       //   // from Element: extension
@@ -21290,35 +21262,20 @@ function addSDCImportFns(ns) {
               var value = [];
 
               for (var j = 0, jLen = answer.length; j < jLen; j++) {
-                var coding = answer[j]; // a valueCoding, which is one of the answers
+                var val = this._processCWECNEValue(answer[j]);
 
-                if (coding.valueCoding) {
-                  value.push({
-                    "code": coding.valueCoding.code,
-                    "text": coding.valueCoding.display
-                  });
-                } // a valueString, which is a user supplied value that is not in the answers
-                else if (coding.valueString) {
-                    value.push({
-                      "text": coding.valueString
-                    });
-                  }
+                if (val) {
+                  value.push(val);
+                }
               }
 
               item.value = value;
             } else {
-              // a valueCoding, which is one of the answers
-              if (qrValue.valueCoding) {
-                item.value = {
-                  "code": qrValue.valueCoding.code,
-                  "text": qrValue.valueCoding.display
-                };
-              } // a valueString, which is a user supplied value that is not in the answers
-              else if (qrValue.valueString) {
-                  item.value = {
-                    "text": qrValue.valueString
-                  };
-                }
+              var val = this._processCWECNEValue(qrValue);
+
+              if (val) {
+                item.value = val;
+              }
             }
 
             break;
@@ -21338,6 +21295,30 @@ function addSDCImportFns(ns) {
             item.value = qrValue.valueString;
         }
       }
+    },
+
+    /**
+     * Handle the item.value in QuestionnaireResponse for CWE/CNE typed items
+     * @param qrItemValue a value of item in QuestionnaireResponse
+     * @returns {{code: *, text: *}}
+     * @private
+     */
+    _processCWECNEValue: function _processCWECNEValue(qrItemValue) {
+      var retValue; // a valueCoding, which is one of the answers
+
+      if (qrItemValue.valueCoding) {
+        retValue = {
+          "code": qrItemValue.valueCoding.code,
+          "text": qrItemValue.valueCoding.display
+        };
+      } // a valueString, which is a user supplied value that is not in the answers
+      else if (qrItemValue.valueString) {
+          retValue = {
+            "text": qrItemValue.valueString
+          };
+        }
+
+      return retValue;
     }
   };
 }